Moves methods out of controller into helper module
